ListView added

+++ YtMultimediaLibrary/ServiceWindow.xaml.cs
@@ -27,7 +27,6 @@ namespace YtMultimediaLibrary
         private readonly YoutubeAPIClient _yt;
         private readonly DataBaseContext _dbContext;
 
-
         public ServiceWindow(User user, UserManager manager, YoutubeAPIClient yt, DataBaseContext dbContext)
         {
             InitializeComponent();
@@ -35,26 +34,30 @@ namespace YtMultimediaLibrary
             _manager = manager;
             _yt = yt;
             _dbContext = dbContext;
-            RefreshDataGrid();
+            RefreshView();
+            LoginLabel.Content = "Signed as: " + _user.UserName;
 
-
-            //_manager.AddUserChannel(user, "https://www.youtube.com/channel/UCSwtGkvmxXhWe-kK1dlm8gA", false);
-            //_manager.AddUserChannel(user, "https://www.youtube.com/channel/UC2bNJW1mb0VaN0-o-llWwAw", false);
-
-            //var channelAndVideos = new Dictionary<Entities.Channel, List<Video>>();
-
-            //foreach (var channel in _user.Channels)
-            //{
-            //    var videos = _yt.ChannelLastVideos(channel, 5);
-            //    //ImageSource img = yt.Foo(videos.First());
-            //    //imgDynamic.Source = img;
-            //    channelAndVideos.Add(channel, videos);
-            //}
-            //lista.ItemsSource = channelAndVideos;
-
+            var channels = _dbContext.Channels.Where(x => x.UserId == _user.UserId).ToList();
+            ChannelsListView.ItemsSource = channels;
         }
 
         
+        private void RefreshView()
+        {
+            foreach (var channel in _dbContext.Channels.Where(x => x.UserId == _user.UserId))
+                _manager.AddUserChannel(_user, channel.Link, false);
+
+            var channelAndVideos = new Dictionary<Entities.Channel, List<Video>>();
+
+            foreach (var channel in _user.Channels)
+            {
+                var videos = _yt.ChannelLastVideos(channel, 5);
+                //ImageSource img = yt.Foo(videos.First());
+                //imgDynamic.Source = img;
+                channelAndVideos.Add(channel, videos);
+            }
+            lista.ItemsSource = channelAndVideos;
+        }
 
         private void VideoClickable_Click(object sender, RoutedEventArgs e)
         {
@@ -68,7 +71,12 @@ namespace YtMultimediaLibrary
         {
             string url = ChannelUrlTextBox.Text;
             string name = ChannelNameTextBox.Text;
-                        
+            
+            if(url.Length <= 0 || name.Length <= 0)
+            {
+                MessageBox.Show("Invalid data");
+                return;
+            }
             var channelToAdd = new Channel
             {
                 Link = url,
@@ -79,28 +87,36 @@ namespace YtMultimediaLibrary
             ChannelUrlTextBox.Text = "";
             ChannelNameTextBox.Text = "";
 
+            _manager.AddUserChannel(_user, url, false);
             _dbContext.Channels.Add(channelToAdd);
             _dbContext.SaveChanges();
+            RefreshView();
         }
 
         private void DeleteChannelButton_Click(object sender, RoutedEventArgs e)
         {
+
             int id = int.Parse(ChannelId.Text);
             var channelToDelete = _dbContext.Channels.SingleOrDefault(x => x.ChannelId == id);
+            if(channelToDelete==null)
+            {
+                MessageBox.Show("Invalid ID");
+                return;
+            }
             _dbContext.Channels.Remove(channelToDelete);
             _dbContext.SaveChanges();
+            RefreshView();
         }
 
-        private void DataGrid_SelectionChanged(object sender, SelectionChangedEventArgs e)
+
+        private void LogoutButton_Click(object sender, RoutedEventArgs e)
         {
-
+            MainWindow mainWindow = new MainWindow();
+            Close();
+            mainWindow.Show();
         }
 
-        private void RefreshDataGrid()
-        {
-            //var listOfChannels = _dbContext.Channels.Where(x => x.ChannelId
-            //ChannelGrid.ItemsSource = listOfChannels;
-        }
 
+     
     }
 }
